requirements.txt
asgiref==3.9.1
contourpy==1.3.0
cycler==0.12.1
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
fonttools==4.59.0
importlib_resources==6.5.2
joblib==1.5.1
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pickle-mixin==1.0.2
pillow==11.3.0
psycopg2-binary==2.9.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
zipp==3.23.0
